Split the job finalize and lock helpers out of pg-queue
pg-queue.ts held the queue classes, the row types, the concurrency-group lock
and the whole finalize path in one file. Three modules come out of it, leaving
873 lines where there were 1151:

- job-tables.ts        row shapes for jobs and job_reservations
- job-concurrency-lock.ts  the advisory lock every writer of a job's outcome takes
- job-finalize.ts      the finalize transaction, its retry policy and outcomes

The split is acyclic on purpose: each new module depends on job-tables at most,
never back on pg-queue, which is why the row types moved rather than being
imported in reverse. The barrel re-exports all three, and nothing imported
pg-queue directly, so no consumer changes.

Also reunite two comments with the declarations they describe — the
serialization-failure predicate had landed between the superseded-cause type
and its own explanation.
