+++ packages/postgres/index.ts
@@ -1,4 +1,7 @@
 export * from './pg-adapter.ts';
 export * from './pg-config.ts';
+export * from './job-concurrency-lock.ts';
+export * from './job-tables.ts';
+export * from './job-finalize.ts';
 export * from './pg-queue.ts';
 export * from './pg-transaction-manager.ts';
